requirements.txt
streamlit==1.25.0
pymupdf==1.22.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.4
langchain==0.2.17
openai==1.40.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
tqdm==4.66.1
gunicorn==20.1.0